v44 - Fix problem with fetching data from Github
It's a problem with the ServiceWorker and the credentials. Per default Serviceworker is not sending
the credentials to GitHub. You need to update the Request that you send to GitHub with the option
credentials: include. But then the Browser throws an error because of CORS. This is while GitHub
allows CORS with a * but not explicit with the requesting website and this is not allowed when
sending credentials.

So as solution for the problem we can temporary no longer cache requests to GitHub with using the
ServiceWorker.

Closes #7